Change PendingXp to u64

+++ src/game/resources/pending_xp_list.rs
@@ -2,14 +2,14 @@ use legion::Entity;
 
 pub struct PendingXp {
     pub entity: Entity,
-    pub xp: i32,
+    pub xp: u64,
     pub source: Option<Entity>,
 }
 
 pub type PendingXpList = Vec<PendingXp>;
 
 impl PendingXp {
-    pub fn new(entity: Entity, xp: i32, source: Option<Entity>) -> Self {
+    pub fn new(entity: Entity, xp: u64, source: Option<Entity>) -> Self {
         Self { entity, xp, source }
     }
 }
